backend/pyproject.toml
[project]
name = "twitter-agent"
version = "0.1.0"
description = "CLI agent for analyzing Twitter users' voice and generating content proposals"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.9.0",
    "httpx>=0.25.0",
    "loguru>=0.7.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dateutil>=2.8.2",
    "pyyaml>=6.0.1",
    "rich>=13.0.0",
    "perplexityai>=0.19.0",
    "readability-lxml>=0.8.1",
    "lxml>=5.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
]

[project.scripts]
twitter-agent = "twitter_agent.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
